+++ example/outlet/lib/app/app_page.dart
@@ -1,4 +1,4 @@
-import 'package:example/routes.dart';
+import 'package:example/routes.g.dart';
 import 'package:flutter/material.dart';
 import 'package:routefly/routefly.dart';