+++ example/simple/lib/app/app_widget.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:routefly/routefly.dart';
 
-import '../routes.dart';
+import '../routes.g.dart';
 
 class AppWidget extends StatelessWidget {
   const AppWidget({super.key});